feat: use unsigned-varint instead of varivari
This commit also includes a lot of small rewriting as, at this stage we
have to figure out what we should do as we develop. In this commit, we
introduced `unsigned-varint` as a substitute for `varivari`, because
varivari had a rather messy codebase and we were struggling to get the
logic right. We also created `RawPacket` to better express the intent of
code inside `handle_connection`. Finally, we changed the default log
level for tracing's FmtSubscriber and reviewed log level in various
locations.

Co-authored-by: tenzyu <tenzyu.on@gmail.com>